migration bug fixed

+++ db/migrate/20140627053118_create_premix_ingredients.rb
@@ -4,7 +4,7 @@ class CreatePremixIngredients < ActiveRecord::Migration
 	    t.belongs_to :formula, null: false
       t.belongs_to :ingredient, null: false
       t.decimal :actual_usage, precision: 12, scale: 6
-      t.integer :premix_usage, precision: 12, scale: 6, default: 0
+      t.decimal :premix_usage, precision: 12, scale: 6, default: 0
       t.timestamps
     end
     add_index :premix_ingredients, [:ingredient_id, :formula_id], unique: true
